fix : just a quick fix forgot to add the async to the new graph
