+++ backend/app/crew.py
@@ -6,8 +6,7 @@ from pathlib import Path
 
 from app.agents.job_requirement_analyst import JobRequirementAnalyst
 from app.agents.search_agent import SearchAgent
-from app.agents.langgraph_react_agent import JobScrutinizerLangGraphAgent
-from app.agents.job_scrutinizer_langgraph import JobScrutinizerLangGraph
+from app.agents.job_scrutinizer_agent import JobScrutinizerLangGraph
 from app.agents.report_generator_agent import json_to_html_table
 from app.tools.mail_sender import send_email
 
@@ -17,7 +16,7 @@ logger = logging.getLogger(__name__)
 
 BASE_DIR = Path(__file__).resolve().parent.parent.parent
 
-# TODO : Get your shit together and fix this crap ASAP
+
 
 async def initialize_crew(user_input_data : dict):
 
@@ -31,9 +30,7 @@ async def initialize_crew(user_input_data : dict):
 
     job_analyst_agent_instance = JobRequirementAnalyst(input= user_input_data , user_id= id)
     search_agent_instance = SearchAgent(user_id= id)
-
-    # job_scrutinizer_agent2 = JobScrutinizerLangGraphAgent(user_id=id , user_input= user_input_data)
-    job_scrutinizer_agent2   = JobScrutinizerLangGraph(user_id=id , user_input= user_input_data)
+    job_scrutinizer_agent   = JobScrutinizerLangGraph(user_id=id , user_input= user_input_data)
 
     logger.info("All agents initialized successfully")
 
@@ -69,7 +66,7 @@ async def initialize_crew(user_input_data : dict):
 
         if results.raw:
 
-            agent_3_result = job_scrutinizer_agent2.scrutinize_jobs()
+            agent_3_result = await job_scrutinizer_agent.scrutinize_jobs()
 
         logger.info("Crew execution completed")
 
@@ -92,7 +89,7 @@ async def initialize_crew(user_input_data : dict):
     except Exception as e:
 
         logging.error(f"The crew Failed miserably bruhhhh : {e}")
-        if len(job_scrutinizer_agent2.saved_jobs) == 0:
+        if len(job_scrutinizer_agent.saved_jobs) == 0:
             send_email(to_email=email, user_id=id,error= True , jobs=0)
             logger.info(f"0 Jobs Email sent successfully to {email}")
             return False
